Use Fredoka One font for minigame text.
This was unusually hard to figure out because of the complexity involved
in determining when a font is actually loaded. Waiting for the page's
'onload' event by itself doesn't seem to work reliably; we need to actually
have initial page content that *uses* the font too. At least, that's what
seems to reliably work right now.

The annoying thing in our case is that it's not just a FOUC if the font
isn't loaded by the time the game starts; text is in Times New
Roman for the duration of the whole game!

In any case, this means that we need to have text on exported games
(and phaser-frame.html) that uses Fredoka One, even if the text is
invisible. For now, that text is being provided by a "byline" at
the bottom left that says "This game was made with MMM".